refacto: make beatLevel to external and use calldata (#2)
Knowing that only the owner can call the function, I think the function
can use the external modifier. Thanks to this, knowing that the function
is external, we can directly use the userData argument from the calldata
layer to save some gas because it is read-only (in this case, we bypass
the implicit conversion to the memory layer).

Additional context:
When the functions get called externally, the arguments are kept in calldata
and copied to memory during ABI decoding (using the opcode calldataload and mstore).
When the argument is read in the function, the value is accesses in memory
using a mload. We can avoid all this logic by simply using calldata for
read-only arguments. In my proposal, instead of going via memory,
the value is directly read from calldata using calldataload.